rename internal to lib for vercel funcs
