+++ api/cmd/server/main.go
@@ -3,9 +3,9 @@ package main
 import (
 	"log"
 
-	"github.com/paulvinueza30/league-portfolio/api/internal/api"
-	"github.com/paulvinueza30/league-portfolio/api/internal/config"
-	_ "github.com/paulvinueza30/league-portfolio/api/internal/services"
+	"github.com/paulvinueza30/league-portfolio/api/lib/api"
+	"github.com/paulvinueza30/league-portfolio/api/lib/config"
+	_ "github.com/paulvinueza30/league-portfolio/api/lib/services"
 )
 
 func main() {
