+++ api/lib/services/registry.go
@@ -9,8 +9,8 @@ import (
 	"slices"
 	"time"
 
-	"github.com/paulvinueza30/league-portfolio/api/internal/config"
-	"github.com/paulvinueza30/league-portfolio/api/internal/models"
+	"github.com/paulvinueza30/league-portfolio/api/lib/config"
+	"github.com/paulvinueza30/league-portfolio/api/lib/models"
 	"github.com/redis/go-redis/v9"
 )
