+++ vercel.json
@@ -3,5 +3,3 @@
     { "source": "/api/(.*)", "destination": "/api/index.go" }
   ]
 }
-
-
